Version 3.0.1

+++ Gury3d/GuiList.cpp
@@ -20,7 +20,7 @@ void RBX::Gui::GuiList::render(RenderDevice* d)
 		end = (pos + size);
 	}
 
-	for (size_t i = 0; i < children.size(); i++)
+	for (unsigned int i = 0; i < children.size(); i++)
 	{
 		GuiObject* c = children.at(i);
 		Vector2 _pos = pos;
